research_data/agent.py
import os
import asyncio
from dotenv import load_dotenv
from openai import OpenAI
from backend.agent_base import BaseAgent, AgentInput, AgentOutput
from backend.research_service import ResearchService

load_dotenv()
client = OpenAI(api_key=os.getenv("OPENAI_API_KEY"))

class ResearchDataAgent(BaseAgent):
    def __init__(self):
        super().__init__()
        self.name = "Research & Data Agent"

    def get_input_keys(self) -> list:
        return ["content_roadmap", "campaign_theme"]

    def get_output_keys(self) -> list:
        return ["research_summary", "trending_topics", "statistics"]

    def run(self, input_data: AgentInput) -> AgentOutput:
        try:
            content_roadmap = input_data.get("content_roadmap", "")
            campaign_theme = input_data.get("campaign_theme", "")
            
            # Extract research query from inputs
            if content_roadmap:
                research_query = f"{campaign_theme} {content_roadmap}"
            else:
                research_query = input_data.get("text", campaign_theme)

            # Perform research using the research service
            research_results = asyncio.run(self._perform_research(research_query))
            
            # Use OpenAI to analyze and summarize the research
            analysis_prompt = f"""
            Based on the following research data about "{campaign_theme}", provide:
            1. A comprehensive research summary
            2. Key trending topics and themes
            3. Important statistics and data points
            4. Actionable insights for content creation

            Research Data:
            {research_results}

            Please structure your response with clear sections and bullet points.
            """

            response = client.chat.completions.create(
                model="gpt-4",
                messages=[
                    {"role": "system", "content": "You are a research analyst specializing in data synthesis and trend analysis."},
                    {"role": "user", "content": analysis_prompt}
                ],
                temperature=0.7,
                max_tokens=1500,
            )

            analysis_content = response.choices[0].message.content

            # Extract structured data from the analysis
            research_summary = self._extract_research_summary(analysis_content)
            trending_topics = self._extract_trending_topics(research_results)
            statistics = self._extract_statistics(research_results)

            return AgentOutput.from_dict({
                "research_summary": research_summary,
                "trending_topics": trending_topics,
                "statistics": statistics,
                "raw_research_data": research_results,
                "analysis": analysis_content,
                "status": "completed",
                "agent": "Research & Data Agent"
            })

        except Exception as e:
            return AgentOutput.from_dict({
                "research_summary": f"[ERROR] Research failed: {str(e)}",
                "trending_topics": [],
                "statistics": {},
                "status": "error",
                "agent": "Research & Data Agent",
                "error": str(e)
            })

    async def _perform_research(self, query: str) -> dict:
        """Perform research using the research service"""
        try:
            async with ResearchService() as research_service:
                filters = {
                    'sources': ['academic', 'web', 'statistics'],
                    'min_relevance': 0.6
                }
                results = await research_service.search(query, filters)
                return results
        except Exception as e:
            # Fallback to simulated research data
            return {
                'query': query,
                'total_results': 5,
                'results': [
                    {
                        'title': f'Research Study: {query}',
                        'content': f'Comprehensive analysis of {query} trends and implications.',
                        'source': 'Academic Database',
                        'relevance_score': 0.9,
                        'data_type': 'academic',
                        'metadata': {'publication_date': '2024-01-15'}
                    },
                    {
                        'title': f'Market Report: {query}',
                        'content': f'Industry insights and market data related to {query}.',
                        'source': 'Market Research',
                        'relevance_score': 0.8,
                        'data_type': 'market-data',
                        'metadata': {'market_size': '$1.2B', 'growth_rate': '12%'}
                    }
                ]
            }

    def _extract_research_summary(self, analysis_content: str) -> str:
        """Extract research summary from analysis"""
        lines = analysis_content.split('\n')
        summary_lines = []
        in_summary = False
        
        for line in lines:
            if 'research summary' in line.lower() or 'summary' in line.lower():
                in_summary = True
                continue
            elif in_summary and ('trending' in line.lower() or 'statistics' in line.lower()):
                break
            elif in_summary and line.strip():
                summary_lines.append(line.strip())
        
        return '\n'.join(summary_lines) if summary_lines else analysis_content[:500]

    def _extract_trending_topics(self, research_results: dict) -> list:
        """Extract trending topics from research results"""
        topics = []
        
        for result in research_results.get('results', []):
            # Extract keywords from titles and content
            title_words = result.get('title', '').split()
            content_words = result.get('content', '').split()[:20]  # First 20 words
            
            # Simple keyword extraction (in real implementation, use NLP)
            keywords = [word.strip('.,!?') for word in title_words + content_words 
                       if len(word) > 4 and word.isalpha()]
            topics.extend(keywords[:3])  # Top 3 keywords per result
        
        # Return unique topics, sorted by frequency
        from collections import Counter
        topic_counts = Counter(topics)
        return [topic for topic, count in topic_counts.most_common(10)]

    def _extract_statistics(self, research_results: dict) -> dict:
        """Extract statistics from research results"""
        statistics = {
            'total_sources_searched': len(research_results.get('sources_searched', [])),
            'total_results_found': research_results.get('total_results', 0),
            'average_relevance_score': 0.0,
            'data_types': {},
            'publication_dates': {}
        }
        
        results = research_results.get('results', [])
        if results:
            # Calculate average relevance
            relevance_scores = [r.get('relevance_score', 0) for r in results]
            statistics['average_relevance_score'] = sum(relevance_scores) / len(relevance_scores)
            
            # Count data types
            for result in results:
                data_type = result.get('data_type', 'unknown')
                statistics['data_types'][data_type] = statistics['data_types'].get(data_type, 0) + 1
            
            # Count publication years
            for result in results:
                pub_date = result.get('metadata', {}).get('publication_date', '')
                if pub_date:
                    year = pub_date[:4]
                    statistics['publication_dates'][year] = statistics['publication_dates'].get(year, 0) + 1
        
        return statistics